adding code which places 10 lines into the hidden area,
